+++ a2/part2/com/RMITGP/examples/symbolic/SymbolicFitness.h
@@ -20,8 +20,8 @@ class SymbolicFitness : public Fitness
    private:
       static const int FITNESS_CASES;
 
-      int *xValues;
-      int *targetFunc;
+      double *xValues;
+      double *targetFunc;
 
    protected:
